Ability Electrocute spell completed.

+++ Source/Aura/Public/UI/WidgetController/AuraWidgetController.h
@@ -37,7 +37,7 @@ struct FWidgetControllerParams	// Facilitating the transfer of parameters
 };
 
 
-DECLARE_DYNAMIC_MULTICAST_DELEGATE_OneParam(FOnPlayerStatChangedSignature, int32, NewValue);
+DECLARE_DYNAMIC_MULTICAST_DELEGATE_OneParam(FOnPlayerStatChangedSignature, float, NewValue);
 DECLARE_DYNAMIC_MULTICAST_DELEGATE_OneParam(FAbilityInfoSignature, const FAuraAbilityInfo&, Info);
 
 UCLASS()
